note/users/views/users.py
"""Users views."""

# Django REST framework
from rest_framework import mixins, status, viewsets
from rest_framework.decorators import action
from rest_framework.response import Response

# Models
from note.users.models import User
from note.notes.models import Note
