requirements/base.txt
# Base
Pillow==8.3.1
psycopg2==2.9.1
pycparser==2.20
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
asgiref==3.4.1
cffi==1.14.6

# Password security
argon2-cffi==20.1.0
# Django
Django==3.2.5

# Enviroment
django-environ==0.4.5

# Django REST Framework
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4

# Static files
whitenoise==5.3.0
